add user authentication to http REST endpoints

+++ src/main/java/com/tvtracker/enterprise/TvTrackerController.java
@@ -41,7 +41,7 @@ public class TvTrackerController {
      * @return a valid user token for session authentication
      */
     @PostMapping(value="/signUp", consumes="application/json", produces="application/json")
-    public ResponseEntity signUpUser(@RequestBody UserAccount userAccount){
+    public ResponseEntity signUpUser(@RequestBody UserAccount userAccount) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
@@ -66,7 +66,7 @@ public class TvTrackerController {
      * @return a valid user token for session authentication
      */
     @GetMapping("/authenticate/{username}/{password}")
-    public ResponseEntity authenticateUser(@PathVariable("username") String username, @PathVariable("password") String password){
+    public ResponseEntity authenticateUser(@PathVariable("username") String username, @PathVariable("password") String password) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
@@ -83,13 +83,24 @@ public class TvTrackerController {
     /**
      * Returns all media entries tied to the given username
      *
+     * 201: successfully updated media entry.
+     * 401: authentication token is invalid.
+     *
      * @param username String uniquely identifying a user
      * @return List user's media entries
      */
-    @GetMapping("/getMediaEntries/{username}")
-    @ResponseBody
-    public List<MediaEntry> getUsersMediaEntries(@PathVariable("username") String username) {
-        return mediaEntryService.fetchMediaEntriesByUsername(username);
+    @GetMapping("/getMediaEntries/{username}/{token}")
+    public ResponseEntity getUsersMediaEntries(@PathVariable("username") String username, @PathVariable("token") String token) {
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+
+        // authenticate request
+        if(!userAccountService.isTokenValid(token, username)) {
+            return new ResponseEntity(headers, HttpStatus.UNAUTHORIZED);
+        }
+
+        List<MediaEntry> mediaEntries = mediaEntryService.fetchMediaEntriesByUsername(username);
+        return new ResponseEntity(mediaEntries, headers, HttpStatus.OK);
     }
 
     /**
@@ -98,18 +109,24 @@ public class TvTrackerController {
      * Returns one of the following status codes:
      * 201: successfully updated media entry.
      * 400: failed to update media entry.
+     * 401: authentication token is invalid.
      *
      * @param mediaEntry a JSON representation of a MediaEntry object
      * @return HttpStatus
      */
-    @PutMapping(value="/editMediaEntry", consumes="application/json", produces="application/json")
-    public ResponseEntity editMediaEntry(@RequestBody MediaEntry mediaEntry){
+    @PutMapping(value="/editMediaEntry/{username}/{token}", consumes="application/json", produces="application/json")
+    public ResponseEntity editMediaEntry(@RequestBody MediaEntry mediaEntry, @PathVariable("username") String username, @PathVariable("token") String token) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
+        // authenticate request
+        if(!userAccountService.isTokenValid(token, username)) {
+            return new ResponseEntity(headers, HttpStatus.UNAUTHORIZED);
+        }
+
         boolean success = mediaEntryService.updateMediaEntry(mediaEntry);
 
-        if(success){
+        if (success) {
             return new ResponseEntity(headers, HttpStatus.OK);
         } else {
             return new ResponseEntity(headers, HttpStatus.BAD_REQUEST);
@@ -122,18 +139,24 @@ public class TvTrackerController {
      * Returns one of the following status codes:
      * 201: successfully deleted media entry.
      * 400: failed to delete media entry.
+     * 401: authentication token is invalid.
      *
      * @param entryId integer uniquely identifying the media entry record
      * @return HttpStatus
      */
-    @DeleteMapping("/removeMediaEntry/{entryId}")
-    public ResponseEntity removeMediaEntry(@PathVariable("entryId") int entryId){
+    @DeleteMapping("/removeMediaEntry/{entryId}/{username}/{token}")
+    public ResponseEntity removeMediaEntry(@PathVariable("entryId") int entryId, @PathVariable("username") String username, @PathVariable("token") String token) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
+        // authenticate request
+        if(!userAccountService.isTokenValid(token, username)) {
+            return new ResponseEntity(headers, HttpStatus.UNAUTHORIZED);
+        }
+
         boolean success = mediaEntryService.deleteMediaEntry(entryId);
 
-        if(success){
+        if (success) {
             return new ResponseEntity(headers, HttpStatus.OK);
         } else {
             return new ResponseEntity(headers, HttpStatus.BAD_REQUEST);
@@ -146,18 +169,24 @@ public class TvTrackerController {
      * Returns one of the following status codes:
      * 201: successfully created media entry.
      * 400: failed to create media entry.
+     * 401: authentication token is invalid.
      *
      * @param mediaEntry a JSON representation of a MediaEntry object
      * @return HttpStatus
      */
-    @PostMapping(value="/addMediaEntry", consumes="application/json", produces="application/json")
-    public ResponseEntity addMediaEntry(@RequestBody MediaEntry mediaEntry){
+    @PostMapping(value="/addMediaEntry/{username}/{token}", consumes="application/json", produces="application/json")
+    public ResponseEntity addMediaEntry(@RequestBody MediaEntry mediaEntry, @PathVariable("username") String username, @PathVariable("token") String token) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
 
+        // authenticate request
+        if(!userAccountService.isTokenValid(token, username)) {
+            return new ResponseEntity(headers, HttpStatus.UNAUTHORIZED);
+        }
+
         boolean success = mediaEntryService.createMediaEntry(mediaEntry);
 
-        if(success){
+        if (success) {
             return new ResponseEntity(headers, HttpStatus.OK);
         } else {
             return new ResponseEntity(headers, HttpStatus.BAD_REQUEST);
